requirements.txt
asgiref==3.5.2
attrs==22.1.0
colorama==0.4.6
Django==3.2.16
djangorestframework==3.14.0
drf-spectacular==0.24.2
exceptiongroup==1.0.4
inflection==0.5.1
iniconfig==1.1.1
jsonschema==4.17.3
packaging==22.0
pluggy==1.0.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyrsistent==0.19.2
pytest==7.2.0
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
